use standard name for plugin default methods class
